fix(formBuilder): Question Matrix settings crash DEV-39 (#5646)
### 📣 Summary

Opening settings for Question Matrix is possible again.

### 💭 Notes

I suspect the thing that broke it was #4403 and it was being broken for
the last year. The investigation into what should appear in the settings
lead me to believe that the `appearance` for `kobomatrix` is always
`'field-list'`, so without much refactor I fixed the problem.

The things I've changed in the code:
- The list of `appearance` types now has a list for `kobomatrix`
- In the `afterRender` callback of `appearance` I fixed how we select
current option, as there was a separate bug in there. In old code
appearance type was always a `string` (type name), and in new code it
can be also `string[]` (with first element being the type name), so we
can't do a simple comparison anymore.

### 👀 Preview steps

Bug template:
1. Open Form Builder
2. Add new row "Question Matrix"
3. Open "Settings" (top right "cog" icon button)
4. 🔴 [on main] notice that settings don't show up and there is error in
console
5. 🟢 [on PR] notice that settings do open correctly