fix(stripe): annotate methods requiring stripe functionality (#5649)
### 📣 Summary
Fixes a bug that was preventing startup when stripe was disabled.


### 💭 Notes
Adds a `stripe_required` annotation to make it clearer which methods
require stripe to be enabled. Also removes an unused method, sets
billing dates to the first of the month when stripe is disabled, and
moves all stripe-disabled functionality to
get_organization_effective_limits, since that's the main method that
other apps interact with.


### 👀 Preview steps
Build KPI with stripe disabled.