feat(massEmails): Celery task for sending emails TASK-1692 (#5659)
### 📣 Summary
Adds a new task to sync with the recurring emails feature. For now this
task is not enabled but it will replace the old task in the near future.


### 👀 Preview steps
For now unit tests cover the functionality, though the task can be
executed manually. These steps may be followed:

1. Set the django emails backend setting to file backend or console
backend in order to be able to debug
2. Set up one or more mass email configs
3. You can use the inactive user query, and change temporarily the
inactive days parameter so that the users email records list is not
empty
4. Enter to the manage.py shell and import the tasks
generate_mass_email_user_lists and _send_emails
5. Execute generate_mass_email_user_lists manually and then _send_emails
6. Review that the emails were sent