ci: pytest with stripe disabled (#5650)
### 💭 Notes
Adds a new GitHub action for running tests with Stripe disabled. In
addition to turning off the boolean setting STRIPE_ENABLED, this
settings file removes stripe apps from INSTALLED_APPS so the
models/fields do not end up in the database. This should expose any
instances where code was relying on these models or fields without an
enclosing `if STRIPE_ENABLED` block.

Also uses `pytest.mark.skipif` to toggle tests that specifically require
stripe to be enabled/disabled. This is better than using
@override_settings because that only flips the STRIPE_ENABLED boolean
and does not affect which models are in the database, so it's possible
for bugs of the type mentioned before to slip through. To make this
work, we needed to take some stripe-specific code out of
`test_data.json`. It was only used by a few test suites however, so it
was just a matter of moving the key generation to a mixin.

The workflow file was just straight up copy-pasted with the command
altered to use the non-stripe environment. Just changing
DJANGO_SETTINGS_MODULE in the job env is not enough because it overrides
it with what is in pyproject.toml.