chore: improve long running migrations system checks message (#5665)
### 📣 Summary
Enhanced the clarity and usefulness of system check messages for
long-running migrations.


### 📖 Description
This update improves the system check messages related to long-running
migrations, making them more descriptive and actionable. The goal is to
help developers and sysadmins quickly understand which migrations are
incomplete and what steps may be required to resolve them.