feat(projectHistoryLogs): add owner to deployment logs DEV-45 (#5647)
### 📣 Summary
Add the project owner to all project history logs dealing with
deployment/archiving.


### 💭 Notes
This PR also tries to set up for easier testing in the future by
conditionally adding the project owner to the method for testing basic
metadata that is called in almost all of the tests.
This PR also tests the number of queries for the updated project history
logs. This is to ensure we aren't accidentally adding queries to
requests to get more information for the project history logs (unless
it's unavoidable). Query counts do tend to vary, so if we end up
constantly having to update the counts whenever we make changes we can
remove it.
The query counts were based on the counts before the project owner was
added.


### 👀 Preview steps

1. ℹ️ have an account
2. Create a new project
3. Deploy the project
4. Redeploy the project
5. Archive the project
6. Unarchive the project
7. Go to `api/v2/<asset_uid>/history`
8. 🟢 There should be four new project history logs, one for each action,
and they should each have your username in the metadata under
`product_owner.`