perf(mongoSync): speed up mongo sync  (#5728)
### 📣 Summary
Improve performance of `sync_mongo` job.


### 💭 Notes
One major inefficiency in the sync_mongo job was that every
ParsedInstance needed to dict-ify itself in isolation, which requires
information that can only be obtained by parsing the XForm. The
ParsedInstances had no way of sharing the information from the XForm,
even though it was the same every time, so the job ended up parsing the
XForm again for every single ParsedInstance, which is very slow for
large XForms.

This PR allows creating the dict version of a submission to be done
outside the ParsedInstance.save() code, then have the ParsedInstance use
the pre-made dict to create the mongo entry. This means that the code
for creating the dict version can share information not available inside
the ParsedInstance itself. Specifically, we can determine the repeating
groups ahead of time, which is the only reason we actually need to parse
the XForm.

Eventually, this could use a larger refactor to make it easier to debug.
Right now, pi.save() calls a very long chain of methods that goes
through Instance, XFormInstanceParser, and DataDictionary, with
necessary data cached on different objects and sometimes inaccessible to
the calling object. It's conceivable we could get rid of ParsedInstance,
XFormInstanceParser, and DataDictionary altogether and just get the
information we need from XForm and Instance objects. ParsedInstance and
XFormInstanceParser are basically layers on top of Instance, and
DataDictionary is basically a layer on top of XForm.

### 👀 Preview steps

On both the release branch and the PR branch:
1. ℹ️ have account and a large project with at least one repeating group
(see internal
[example](https://chat.kobotoolbox.org/#narrow/stream/4-Kobo-Dev/topic/Inefficency.20of.20ParsedInstance.2Esave.28.29/near/602833))
2. Create about ~10 submissions (should be enough to see the time
difference).
3. Note: if you're using the xls provided, the easiest way to make 10
submissions without having to go through a massive form is to select
'PAP type: MST'. This form has been modified from the one originally
taken from prod. It may also be faster to just create one or two
submissions and then use the `api/v2/assets/uid/data/id/duplicate`
endpoint to copy them as what's in the submission is not terribly
important
4. Run the Django IPython shell
5. `from django.core.management import call_command`
6. `time call_command('sync_mongo', '-a', '-r', <username>, <project
uid>)`
7. :green_circle: the process should be much faster on the PR branch
than on main
8. In the mongo container, run `mongoexport --db=formhub
--collection=instances --query='{"_xform_id_string": "<uid>"}'
--out=/tmp/instances.json` and copy the results locally
9. Once you have copied both .jsons locally, compare them using your
favorite diff tool to make sure they are the same