docs(api): adding API endpoints for access-logs in KPI documentation TASK-1788 (#5722)
### 📣 Summary
Added API documentation for `/api/v2/docs/access-logs` endpoints.

### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the access logs endpoints. With
these changes, users will be able to understand, test and use the access
logs endpoints more easily.

### 💭 Notes
The original comments have been preserved for future use in the Python
documentation.
This PR supersedes #5709

### 👀 Preview steps
To see the changes, visit http://kf.kobo.local/api/v2/docs/. The page of
the Kobo API will now be visible and generated by swagger. To see the
changes added to access-logs endpoint, scroll down to the category of
the same name. Documentation for each endpoint will be provided with a
HTTP code, a body and/or a response when necessary.

---------

Co-authored-by: zoegodin-admin <2133178@cshawi.ca>
Co-authored-by: Olivier Leger <olivierleger@gmail.com>