docs(api): adding API endpoints for asset-snapshots in KPI documentation TASK-1791 (#5739)
### 📣 Summary
Added API documentation for `/api/v2/docs/asset-snapshots` endpoints.


### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the asset-snapshots endpoints. With
these changes, users will be able to understand, test and use the
asset-snapshots endpoints more easily.

### 💭 Notes
The tags for the OpenRosa endpoints have been added so they wouldn't be
included in the asset-snapshots endpoint tag

### 👀 Preview steps
To see the changes, visit `http://kf.kobo.local/api/v2/docs/`. The page
of the Kobo API will now be visible and generated by swagger. To see the
changes added to asset-snapshots endpoint, scroll down to the category
of the same name. Documentation for each endpoint will be provided with
a HTTP code, a body and/or a response when necessary.

---------

Co-authored-by: zoegodin-admin <2133178@cshawi.ca>
Co-authored-by: Olivier Leger <olivierleger@gmail.com>