fix(permissions): make sharing form request permissions upon opening DEV-128 (#5785)
### 📣 Summary
<!-- Delete this section if changes are internal only. -->
<!-- One sentence summary for the public changelog, worded for
non-technical seasoned Kobo users. -->

Fixes issue where wharing modal did not update permissions until a hard
refresh.

### 💭 Notes
<!-- Delete this section if empty. -->
<!-- Anything else useful that's not said above,worded for
reviewers, testers, and future git archaeologist collegues. Examples:
- screenshots, copy-pasted logs, etc.
- what was tried but didn't work,
- conscious short-term vs long-term tradeoffs,
- proactively answer likely questions,
-->
Originally thought of as a caching issue, but that's not the case. The
modal is just not requesting permissions on open.

One liner but that raises some questions. It feels like something we
would have made a decision to not do in the past since it's a simple get
request that would make sense to be there. I tried looking in the commit
history to see if this line was ever removed but didn't find anything 🤷

A better solution would use a store, or react query, but until then this
modal should be updating correctly

### 👀 Preview steps
<!-- Delete this section if behavior can't change. -->
<!-- If behavior changes or merely may change, add a preview of a
minimal happy path. -->

Bug template:
1. ℹ️ have multiple accounts and an asset
2. Open the sharing modal anywhere
3. Grant some permissions to a user (can really be any permission
change, but these steps follow adding a user)
4. Close the modal
5. Open the modal again
6. 🔴 [on main] notice that the permissions are not updated and the new
user is not in the modal
7. 🟢 [on PR] notice that the permissions are updated and the new user is
in the modal on after closing and opening the sharing modal again