fix(submissions)!: reject submissions with identical basename (#5794)
### 📣 Summary
Prevented multiple submissions with the same basename from being
accepted.


### 📖 Description
This fix ensures that identical submissions, those with the same XML
content and the same `meta/instanceID`, are properly rejected if they
are submitted again using the same basename. The response will be **202
- Duplicated submission**.

Previously, a submission could be accepted multiple times under the same
basename if it was retried, leading to inconsistent states and potential
storage issues.