docs(api): adding API endpoints for assets in KPI documentation TASK-1794 (#5783)
### 📣 Summary
Added API documentation for `/api/v2/docs/assets` endpoint.

### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the `assets` endpoint. With these
changes, users will be able to understand, test and use the `assets`
endpoint more easily.


### 👀 Preview steps
To see the changes, visit `http://kf.kobo.local/api/v2/docs/`. The page
of the Kobo API will now be visible and generated by swagger. To see the
changes added to `assets` endpoint, scroll down to the category of the
same name or make a research in the search bar. Documentation for the
endpoint will be provided with a HTTP code, a body and/or a response
when necessary.