feat(massEmails): adapt timeouts for condensed send mode (#5821)
### 📣 Summary
Lower the timeouts for mass email sends when in test mode.

### 💭 Notes
Update the time_limit option in the task decorator to be conditional on
whether or not we're in condensed send mode. Also adds the
soft_time_limit option because celery will ignore the regular time_limit
if soft_time_limit is higher.


### 👀 Preview steps
If you are using kobo-compose, you may need to recreate the kpi_worker
container (not just restart it).
Bug template:
1. ℹ️ have a superuser account and at least one other account
2. Update `MASS_EMAIL_THROTTLE_PER_SECOND` to 1 and
`MASS_EMAIL_SLEEP_SECONDS` to 400
3. Add at least two user emails to `MASS_EMAIL_TEST_EMAILS`
4. Create a new MassEmailConfig with frequency=1 and using the
test_users query. Set it to live.
5. 🟢 At the next 15 min boundary (+1 minute), the send_emails job should
run. It should send one email then throw a
TimeLimitExceeded error after 3 minutes