fix(attachments): hide the "Delete media files only" button from any permission except `change_submissions` or higher DEV-583 (#5838)
### 📣 Summary

Those with `delete_submissions` permission for submissions now cannot
see the "Delete media files only" button in the bulk options and those
with `change_submissions` permission are now able to see the button.

### 💭 Notes

Missed a spot when writing the guards for displaying the button.
Clicking it would trigger another set of guards in the actual component
so the user couldn't delete attachments this way anyway.

### 👀 Preview steps

1. ℹ️ have 3 accounts and a project with media attachment submissions
2. allow a `user_a` to have `delete_submissions` permissions
3. allow a `user_b` to have `change_submissions` permissions
4. 🔴 [on main] notice that `user_a` **can** see "Delete media files
only" button in the bulk select options but `user_b` **cannot**
6. 🟢 [on PR] notice `user_a` **cannot** see "Delete media files only"
button, but `user_b` **now can**