fix(massEmails): convert dates to strings for xslx DEV-601 (#5847)
### 📣 Summary
Fix exporting mass email records to XSLX format.

### 💭 Notes
The XLSX exporter doesn't work with dates with timezones, so convert all
dates to strings before the exporting happens. This way no information
is lost and whoever exports the data can manipulate the dates in Excel
as they wish.


### 👀 Preview steps
Remember to restart the kpi_worker after switching branches.
1. ℹ️ have an account and a MassEmailConfig that has sent at least once
2. Select the config and select the 'Export with celery' action
3. On the ExportJob form, select the XSLX format and the Last 30 days
recipients resource and save.
4. 🔴 [on main] The job hangs. In the kpi_worker logs you will see
`TypeError: Excel does not support timezones in datetimes. The tzinfo in
the datetime/time object must be set to None.`
5. 🟢 [on PR] The export completes and you are able to download the xslx