docs(api): fix schema errors DEV-643 (#5874)
### 📣 Summary
Fixed schema fail validations that were raising.

### 👀 Preview steps
Go to `http://kf.kobo.local/api/v2/docs/`. The page of the Kobo API will
now be visible and generated by swagger. Check in the assets `bulk` and
`patch` payload, when clicking on `schema`, that the notion of `oneOf`
is visible with the different options. Also make sure when typing
`http://kf.kobo.local/api/v2/assets.json?metadata=on` that all fields
are presents.